Rework mobile closure as two-step form
The user is really doing one action here - closing a work order. Having
them as separate pages didn't have much benefit - revisiting the
confirmation page doesn't seem useful. This commit makes this a
two-step form as we do for manager closures. It also means we can
store the overtime state within regular React state rather than in
a query param.
